# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Masuka Junior School ECD API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stories/{id}:
    get:
      operationId: getStory
      tags: [stories]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: A story
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Story" }
...
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      parameters:
        - in: query
          name: assignmentId
          schema: { type: string }
        - in: query
          name: studentId
          schema: { type: string }
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Submission" }
    post:
      operationId: createSubmission
      tags: [submissions]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSubmissionBody" }
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Submission" }
  /submissions/{id}/grade:
    post:
      operationId: gradeSubmission
      tags: [submissions]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GradeSubmissionBody" }
      responses:
        "200":
          description: Graded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Submission" }
...
  /students/{id}/progress:
    get:
      operationId: getStudentProgress
      tags: [students]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Auto-generated progress report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProgressReport" }
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UploadUrlRequest" }
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadUrlResponse" }
        "400":
          description: Invalid input
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
components:
  schemas:
...
    Story:
      type: object
      required: [id, title, emoji, classLevel, summary, body, readMinutes]
      properties:
        id: { type: string }
        title: { type: string }
        emoji: { type: string }
        classLevel: { type: string }
        summary: { type: string }
        body: { type: string }
        moral: { type: string, nullable: true }
        readMinutes: { type: integer }
...
    Attachment:
      type: object
      required: [url, name, mimeType, kind]
      properties:
        url: { type: string }
        name: { type: string }
        mimeType: { type: string }
        kind: { type: string, enum: [image, audio, video, file] }
        size: { type: integer, nullable: true }
    Submission:
      type: object
      required: [id, assignmentId, assignmentTitle, studentId, studentName, submittedAt, content, status, attachments]
      properties:
        id: { type: string }
        assignmentId: { type: string }
        assignmentTitle: { type: string }
        studentId: { type: string }
        studentName: { type: string }
        submittedAt: { type: string, format: date-time }
        content: { type: string }
        attachmentUrl: { type: string, nullable: true }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
        status: { type: string, enum: [pending, graded] }
        grade: { type: string, nullable: true }
        feedback: { type: string, nullable: true }
        gradedAt: { type: string, format: date-time, nullable: true }
    CreateSubmissionBody:
      type: object
      required: [assignmentId, content]
      properties:
        assignmentId: { type: string }
        content: { type: string }
        attachmentUrl: { type: string, nullable: true }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, minLength: 1 }
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
    GradeSubmissionBody:
      type: object
      required: [grade, feedback]
      properties:
        grade: { type: string, enum: [excellent, good, developing, needs_support] }
        feedback: { type: string }
    Student:
      type: object
      required: [id, fullName, classLevel, avatarColor, submissionCount, gradedCount]
      properties:
        id: { type: string }
        fullName: { type: string }
        classLevel: { type: string }
        avatarColor: { type: string }
        submissionCount: { type: integer }
        gradedCount: { type: integer }
    Milestone:
      type: object
      required: [name, area, level]
      properties:
        name: { type: string }
        area: { type: string }
        level: { type: string, enum: [emerging, developing, achieved, excelling] }
    ProgressReport:
      type: object
      required: [student, totalAssignments, submittedCount, gradedCount, gradeBreakdown, milestones, generatedAt, summary]
      properties:
        student: { $ref: "#/components/schemas/Student" }
        totalAssignments: { type: integer }
        submittedCount: { type: integer }
        gradedCount: { type: integer }
        gradeBreakdown:
          type: array
          items:
            type: object
            required: [grade, count]
            properties:
              grade: { type: string }
              count: { type: integer }
        milestones:
          type: array
          items: { $ref: "#/components/schemas/Milestone" }
        generatedAt: { type: string, format: date-time }
        summary: { type: string }
...